backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
django-cors-headers>=4.3
django-ratelimit>=4.1
drf-spectacular>=0.27
bcrypt>=4.1
python-dotenv>=1.0
